Move CbusAllocateNodeNumber to swing class
Improvements to Threading
Increase test delay from 5ms to 10ms
