+++ java/src/jmri/jmrix/can/cbus/node/CbusNodeTableDataModel.java
@@ -7,12 +7,15 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.TimerTask;
+
 import javax.annotation.Nonnull;
+
 import jmri.jmrix.can.*;
 import jmri.jmrix.can.cbus.CbusConstants;
 import jmri.jmrix.can.cbus.CbusMessage;
 import jmri.jmrix.can.cbus.CbusPreferences;
 import jmri.jmrix.can.cbus.CbusSend;
+import jmri.jmrix.can.cbus.swing.nodeconfig.CbusAllocateNodeNumber;
 import jmri.jmrix.can.cbus.swing.nodeconfig.NodeConfigToolPane;
 import jmri.util.*;
